Remove Closeable from BaseStatement
Interestingly enough, there are only very few changes required to make
BaseStatement not implement Closable. And the places are few and all internal.

As we have never actually published that the various SqlStatement
things are closable (and basically no one ever should have to deal
with them directly anyway), it should be possible to make this change
without causing too much scare amongst JDBI users. Or not?

Discussion wanted!