build all test classes natively if supported
QnD attempt to add native compilation to the tree, run with `-Dnative`.

Note: Does not work with the out of the box plugin but requires
https://github.com/graalvm/native-build-tools/pull/720 applied to the
graalvm native plugin and that the embedded junit version (in
https://github.com/graalvm/native-build-tools/blob/master/gradle/libs.versions.toml)
matches the version in jdbi (currently 5.13.0-M2).

native images build but don't work. This is only a "proof of concept, compiles" sketch.